Added all french pages for language switch button
Had to link to french versions of the about and form js files since they insert text dynamically.

This led me to find that I had a duplicate list item in skills section about page pop up that I was able to correct

I need to program the language and theme button to skidaddle on about, work, and contact pages. I for sure want them visible (and functional, of course) on the home page, but also maybe on the nav. That will be tricky programming since the nav exists on all pages, but not in the main. I started to try this, but just saved the commented out code for later