Improvements to form
When testing form submission on mobile device, I realized the alert is no good because Form Submit requires a captcha thingy in order to submit. It has a success notification built in. Removed the alert in my file

The form wasn't submitting during testing, so I recopied changes to fr-form.js