Attempting to prevent resume link from automatic download
In certain browsers, such as Chrome on my laptop, my resume link is triggering the PDF to automatically be downloaded. I want the viewer to have the option to view or at least confirm the download prior to the download, rather than just using a link that auto-downloads. It gives a forceful feel for user experience...

To try to prevent this, I added target=_blank attribute. This worked fine for mobile device (iPhone XR), but did not solve issue on Chrome on PC... Using ChatGPT advice, I added type attribute and and changed download to download=' ' empty string.

There is an additional option to embed a PDF viewer directly in my HTML, but I'm wondering if that will break functionality if device already has a PDF viewer, making more problems where they didn't previously exist. Will try the first two strategies first and return to the inline PDF viewer option if need be