Attempting to fix PDF auto download issue
Adding target= "_blank" and download equals empty string and type = "application/pdf" did not fix the auto PDF download.. To clarify, it functions perfectly well on page opened locally but not live site.

I wanted to try the embed inline PDF viewer but it looks what when I just insert in the body of the about page html, so I tried making it a separate page in a new html.

Going to test it on live site (again) and return to this issue if need be