Changing nav button event listener to target container
This is meant for better user experience, so the target is bigger

I'm noticing that if I change the height of the <i> element or the <li> container, it throws off the icon. This has something to do with the way that font awesome icons work with the :before pseudo element. This is for both nav button and theme button.

I'm going to try this and see if user experience better on testing mobile.. If not I can return to this sizing/height container issue