Moved button containers to higher z-index
Moved to top of stack on z-axis so that event listeners always function properly. Will test on live site

No way to target pseudo elements (icons themselves) since they are not part of the DOM

Wasn't sure if I needed to change second if statement in nav.js (when DOM content loaded) because that doesn't seem to be causing an issue... Theme button and language button are set to display on first load of home page anyway. Hope that makes sense