Corrected theme switch event listener
To amend my previous commit description, that did NOT make sense and I did in fact need to edit my second if statement in theme.js because obviously I need to apply event listener to theme button on first load. ADDITIONALLY, I didn't even need the if statement because the event listener can (dare I say, 'should'?) be applied to the button. I think it's fine to always apply event listener even if the icon(and container) is set to display: none.