Changed files to fix HTML validator errors
Ran through HTML validator again. Role='navigation' attribute no longer needed on nav button now that it's a button?

There were multiple errors on resume.html, but the only one I changed was the download attribute (not allowed on iframe element) I didn't change the height and width attributes because that messes up the way it displays. Leaving height and width as is