Added click event listener to links via JS vs. inline
After finding out that my <a> elements needed an href value in order to be identified and picked up in the tabbing of pages, I furthered researched how to maintain my transitionToPage function while having a defined href value. I edited the transitionToPage function to read the inline href of the elements. Then, I started thinking that I should research whether it was better for accessibility to have inline onclick events trigger the function or if I should assign event listeners via JavaScript. My research suggested that JS is better than inline HTML because it's better for accessibility, allowing users to navigate normally whether JS is enabled or not. It's better semantic HTML practice. And it prevents your HTML from becoming muddled.

Aside from editing nav.js transitionToPage function and adding event listener to all links, I also changed the language button to a link. We can still call it a button and it can be grouped with the other header buttons, but technically it does take the user to another HTML page, so semantically, it's better to use an <a> element. I did decide to omit the target='_blank' attribute because it is the only link that really shouldn't open in another tab