Work pages edit
I decided to take off the placeholder image for now because I already have quite a few projects to showcase and I will be adding more, at which point, I'll have to think about actually picking just a few favorites to showcase on the portfolio site.

I also tried to reword the description of the last project, Off the Hook, in order to better describe that it's a real, pro-bono project for an actual small business, not a fictional one. This came as a suggestion after brother tested site