Added Techdegree certificate link
Linked in education section of about page. Now that I have 2 PDFs, I also changed the download icon next to resume to a PDF icon. I think this better conveys to the user that clicking the icon (button) or the link will allow you to view PDF (then download if desired).