Removed CSS rules for text transform: uppercase
I removed text transform to uppercase for everything except the form (I like it uppercase). I figure if I need to use a different font for readability in paragraphs or certain elements, I can find a different font in the future. Unica One sentence case might be hard to read if used for more than titles or large font elements, but I don't like Varela Round as a secondary font