Removed slide-in effect for footer
Since I copied the footer to all secondary pages, the animation for the footer to slide in would overlap page content when scrolling is necessary on smaller screens. I chose to nix slide in animation for the footer and just have it fade in and out like the main

The footer is now fading in and out properly when navigating and opening/closing nav menu. However, further formatting is required to integrate footer into about, contact, and work pages properly