Nav button icon change
Hamburger icon not appropriate nav on screens wider than mobile. Changed to hamburger icon on mobile and compass icon on wider screens

Reprogrammed nav.js to reflect that. Still need to keep both icons and display based on media queries so that it reappears if resized on desktop

Copied changes to header across pages