Hiding some of the projects
I had a total of 15 or 16 projects and it was cluttered; taking a few off now and seeing how I like having 12. Still a lot--probably should dwindle it down more. Leaving the code in and just commenting out so I can change my  mind as many times as I want. Will likely return to this concept after reviewing and refactoring projects.