Fixed background flicker on initial load
Still have the tiniest flicker of buttons on load/refresh but that is more complex... CSS rules will stay because I'll try to revisit this later. For now, this is much better. Refreshing makes a flicker on light mode because of the standard settings. May have to research more or consult someone on these more complex flicker issues